+++ app/components/reportProblem/__test__/ReportProblem.test.tsx
@@ -1,11 +1,11 @@
 import { fireEvent, render } from "@testing-library/react";
 import type { Survey } from "posthog-js";
 import { ReportProblem } from "~/components/reportProblem/ReportProblem";
-import { fetchSurvey } from "~/services/analytics/fetchSurveys";
+import { fetchSurvey } from "~/services/analytics/surveys/fetchSurveys";
 import { useAnalytics } from "~/services/analytics/useAnalytics";
 
 describe("ReportProblem", () => {
-  vi.mock("~/services/analytics/fetchSurveys");
+  vi.mock("~/services/analytics/surveys/fetchSurveys");
   vi.mock("~/services/analytics/useAnalytics");
 
   vi.mocked(fetchSurvey).mockReturnValue({
